crypto-ai-trader/requirements.txt
# Core Trading Libraries
binance-connector==3.12.0
ccxt==4.1.52
pandas==2.1.3
numpy==1.26.2

# AI/LLM Integration
anthropic==0.20.0
openai==1.3.8

# Data & Analysis
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter
redis==5.0.1

# Monitoring & Logging
python-telegram-bot==19.7
Flask==3.0.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pytz==2023.3
requests-cache==1.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
ipython==8.18.1
jupyter==1.0.0
ipykernel==6.26.0
